+++ inst/examples/ex-phase.R
@@ -1,8 +1,9 @@
 ## Coerce to MCMC objects
 ## Events
-(eve <- as_events(events[1:10000, ], iteration = 1))
+(eve <- as_events(events[1:10000, ], calendar = "CE", iteration = 1))
 ## Phases
-(pha <- as_phases(phases[1:10000, ], start = c(1, 3), iteration = 1))
+(pha <- as_phases(phases[1:10000, ], start = c(1, 3), calendar = "CE",
+                  iteration = 1))
 
 ## Compute phases from events/dates
 ## Compute min-max range for all chains
@@ -10,12 +11,12 @@ pha1 <- phase(eve)
 
 ## Compute min-max range by group
 ## Unless otherwise specified, the phases are assumed to be unordered
-pha2 <- phase(eve, groups = list(A = c(1, 3), B = c(2, 4)))
+pha2 <- phase(eve, groups = list(phase_1 = c(1, 3), phase_2 = c(2, 4)))
 
 all(pha == pha2) # TRUE
 
 ## Set chronological order
 ## (from the oldest to the youngest phase)
-set_order(pha2) <- c("B", "A")
+set_order(pha2) <- c("phase_2", "phase_1")
 get_order(pha2)
 is_ordered(pha2)