+++ inst/examples/ex-test.R
@@ -6,4 +6,4 @@ eve <- eve[1:10000, ]
 older(eve)
 
 ## Test for hiatus
-lapse(eve)
+hiatus(eve[, 2], eve[, 3])